+++ changelog.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [0.3.0] - 2024-06-24
+### Features
+* Fix loading of Form Designer
+* Add cancel button
+* Add exception display 
+* Visual tweaks to progress bar
+
 ## [0.2.0] - 2024-06-21
 ### Features
 * Added reuse of compiled C# expressions from message payload, repeated use of an expression executes faster